Rounding out the code a bit

+++ Testing.Client/Helpers/INumberFormatter.cs
@@ -2,6 +2,6 @@
 {
     public interface INumberFormatter
     {
-        int GetFormattedNumber(string providedNumber);
+        long GetFormattedNumber(string providedNumber);
     }
 }
